backend/pyproject.toml

[tool.poetry]
name = "test-spectr"
version = "0.1.0"
description = "A test task project for Spectr"
authors = ["Andrey Chernikov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.23.1"
django-debug-toolbar = "^3.5.0"
django-filter = "^22.1"
django-phonenumber-field = "^6.3.0"
phonenumbers = "^8.12.53"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
